dealer tests and bugfixes

+++ BuckshotRoulette/itemsHandler.h
@@ -19,7 +19,7 @@ void item_add(ITEM_T *arr, size_t n);
 
 const char *item_text(ITEM_T item);
 
-size_t findItem(ITEM_T item, ITEM_T *items, size_t items_n);
+int findItem(ITEM_T item, ITEM_T *items, size_t items_n);
 
 ITEM_T item_remove(ITEM_T *arr, size_t position);
